Dimensions3-BETA5
- Small changes that were required for the addons to work

- HubWorldAddon
Change the name of the .jar. Make sure to delete the old one to prevent having a duplicate of the addon

- UnbreakableAddon
Change the name of the .jar. Make sure to delete the old one to prevent having a duplicate of the addon

+BiomesAddon
+ChargeOnUseAddon
+DaylightAddon
+LimitedUsesAddon
+MessageOnUse
+PastedPortalsAddon
+PermissionsAddon
+RandomLocationAddon
+TimedPortalsAddon
+WorldGuardAddon

All the newly added addons (except from the Daylight addon) are configured the same way as the old addons, so you can consult the wiki[https://astaspastagam.gitbook.io/first-steps/configuring-dimensions/addons] for info about the configuration.